# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "register.",
                "operationId": "73fd9b61dff4256cc0ee710de66723d9",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name",
                        "schema": {
                            "type": "string",
                            "example": "Anton"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email",
                        "schema": {
                            "type": "string",
                            "example": "anton.fullstack@gmail.com"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "password",
                        "schema": {
                            "type": "string",
                            "example": "password"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "register.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "name": "Anton",
                                                "email": "anton.fullstack@gmail.com",
                                                "updated_at": "2024-02-03T16:53:20.000000Z",
                                                "created_at": "2024-02-03T16:53:20.000000Z"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "login.",
                "operationId": "bb5bdca9f30cb60a33c1ae936a833da2",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email",
                        "schema": {
                            "type": "string",
                            "example": "anton.fullstack@gmail.com"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "password",
                        "schema": {
                            "type": "string",
                            "example": "Tomas1989"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display current token.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "access_token": "Bearer 2|NdjBByPTcx9DzQgSN8x7dq3DUtjDLmEbwNQqvPGEfdb4e5db"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "logout.",
                "operationId": "618f033d5ecfff68b1d0880a5f399d7b",
                "responses": {
                    "401": {
                        "description": "Unauthorized"
                    },
                    "200": {
                        "description": "logout.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Display current user.",
                "operationId": "9c56ef837e374a061a0505742a2a277d",
                "responses": {
                    "401": {
                        "description": "Unauthorized"
                    },
                    "200": {
                        "description": "Display current user.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "name": "Anton",
                                                "email": "anton.fullstack@gmail.com",
                                                "updated_at": "2024-02-03T16:53:20.000000Z",
                                                "created_at": "2024-02-03T16:53:20.000000Z"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/api/remove": {
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "remove current user.",
                "operationId": "c858a10a99962bfe9b26324e20a0f627",
                "responses": {
                    "401": {
                        "description": "Unauthorized"
                    },
                    "200": {
                        "description": "remove current user.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/api/checkout": {
            "post": {
                "tags": [
                    "Checkout"
                ],
                "summary": "create order.",
                "operationId": "e126268a193f1b6458b943b54608e02a",
                "requestBody": {
                    "description": "create order",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "products": {
                                        "title": "products",
                                        "type": "array",
                                        "items": {
                                            "type": "object",
                                            "example": {
                                                "id": 1,
                                                "quantity": 1
                                            }
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthorized"
                    },
                    "200": {
                        "description": "create order.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/api/orders": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Display orders",
                "operationId": "6042ab9e22d46627475a39b3c2d59026",
                "responses": {
                    "401": {
                        "description": "Unauthorized"
                    },
                    "200": {
                        "description": "Display orders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "title": "orders",
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "example": {
                                                    "total_price": 1,
                                                    "created_at": "2024-02-03T16:53:20.000000Z",
                                                    "updated_at": "2024-02-03T16:53:20.000000Z",
                                                    "line_items": [
                                                        {
                                                            "price": 111,
                                                            "quantity": 1
                                                        }
                                                    ]
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Display products",
                "operationId": "55752f3015ba0bb55b2a43348978ac14",
                "responses": {
                    "200": {
                        "description": "Display products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "title": "products",
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "example": {
                                                    "title": "Product",
                                                    "description": "Product description",
                                                    "category": "Product category",
                                                    "updated_at": "2024-02-03T16:53:20.000000Z",
                                                    "created_at": "2024-02-03T16:53:20.000000Z"
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "description": "Api Key for authorization.",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}